+++ .gitignore
@@ -5,5 +5,6 @@ tensorboard/*
 
 !data/cache/.gitkeep
 !data/osm/.gitkeep
+!data/osm/luxembourg-latest.osm.pbf
 !checkpoints/.gitkeep
 !tensorboard/.gitkeep
