--- geobacter/mapnik/.travis.yml
@@ -1,28 +1,0 @@
-os: linux
-language: minimal
-services:
-- docker
-# DOCKER_USERNAME/DOCKER_PASSWORD have been configured in Travis CI
-# env:
-before_install:
-- echo "Before install"
-before_script:
-- docker pull overv/openstreetmap-tile-server || true
-script:
-- docker build --pull --cache-from overv/openstreetmap-tile-server --tag overv/openstreetmap-tile-server .
-- docker volume create openstreetmap-data
-- docker run --rm -v openstreetmap-data:/var/lib/postgresql/12/main overv/openstreetmap-tile-server import
-- docker run --rm -v openstreetmap-data:/var/lib/postgresql/12/main -p 8080:80 -d overv/openstreetmap-tile-server run
-- sleep 30
-- make DOCKER_IMAGE=overv/openstreetmap-tile-server stop
-after_script:
-- docker images
-after_success:
-- if [[ "$TRAVIS_BRANCH" == "master" ]];
-  then
-  docker images ;
-  echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin ;
-  docker push overv/openstreetmap-tile-server ;
-  fi
-notifications:
-  email: true
