--- geobacter/mapnik/apache.conf
@@ -1,17 +1,0 @@
-<VirtualHost *:80>
-    ServerAdmin webmaster@localhost
-
-    LoadTileConfigFile /usr/local/etc/renderd.conf
-    ModTileRenderdSocketName /var/run/renderd/renderd.sock
-    ModTileRequestTimeout 0
-    ModTileMissingRequestTimeout 30
-
-    DocumentRoot /var/www/html
-
-    ErrorLog ${APACHE_LOG_DIR}/error.log
-    CustomLog ${APACHE_LOG_DIR}/access.log combined
-
-    <IfDefine ALLOW_CORS>
-        Header set Access-Control-Allow-Origin "*"
-    </IfDefine>
-</VirtualHost>
