--- geobacter/mapnik/postgresql.custom.conf.tmpl
@@ -1,25 +1,0 @@
-# Suggested minimal settings from
-# https://ircama.github.io/osm-carto-tutorials/tile-server-ubuntu/
-
-shared_buffers = 128MB
-min_wal_size = 1GB
-max_wal_size = 2GB
-maintenance_work_mem = 256MB
-
-# Suggested settings from
-# https://github.com/openstreetmap/chef/blob/master/roles/tile.rb#L38-L45
-
-max_connections = 250
-temp_buffers = 32MB
-work_mem = 128MB
-wal_buffers = 1024kB
-wal_writer_delay = 500ms
-commit_delay = 10000
-# checkpoint_segments = 60 # unrecognized in psql 10.7.1
-max_wal_size = 2880MB
-random_page_cost = 1.1
-track_activity_query_size = 16384
-autovacuum_vacuum_scale_factor = 0.05
-autovacuum_analyze_scale_factor = 0.02
-
-listen_addresses = '*'
